feat: 엔티티에 DiscountType 컬럼 추가 및 수정

+++ backend/main-service/src/main/java/kkakka/mainservice/coupon/application/DiscountService.java
@@ -6,6 +6,7 @@ import kkakka.mainservice.category.domain.Category;
 import kkakka.mainservice.category.domain.repository.CategoryRepository;
 import kkakka.mainservice.common.exception.KkaKkaException;
 import kkakka.mainservice.coupon.domain.Discount;
+import kkakka.mainservice.coupon.domain.DiscountType;
 import kkakka.mainservice.coupon.domain.repository.DiscountRepository;
 import kkakka.mainservice.coupon.ui.dto.DiscountRequestDto;
 import kkakka.mainservice.coupon.ui.dto.DiscountResponseDto;
@@ -27,37 +28,37 @@ public class DiscountService {
     /* 할인 등록 */
     @Transactional
     public Long createDiscount(DiscountRequestDto discountRequestDto) {
-        /* 상품 할인 */
-        if (discountRequestDto.getCategoryId() == null) {
-            Product product = getProduct(discountRequestDto);
-            Discount discount = Discount.create(
-                product,
-                discountRequestDto.getName(),
-                discountRequestDto.getDiscount(),
-                discountRequestDto.getStartedAt(),
-                discountRequestDto.getExpiredAt());
-            product.changeDiscount(discountRequestDto.getDiscount());
-            productRepository.save(product);
-            discountRepository.save(discount);
-            return discount.getId();
-        }
-
-        /* 카테고리 할인 */
-        else if (discountRequestDto.getProductId() == null) {
-            Category category = getCategory(discountRequestDto);
-            Discount discount = Discount.create(
-                category,
-                discountRequestDto.getName(),
-                discountRequestDto.getDiscount(),
-                discountRequestDto.getStartedAt(),
-                discountRequestDto.getExpiredAt());
-            List<Product> products = categoryRepository.findProductsByCategoryId(category.getId());
-            for (Product product : products) {
+        if (discountRequestDto.isValidateDate() && discountRequestDto.isValidateDiscount()) {
+            if (DiscountType.PRODUCT_DISCOUNT.equals(discountRequestDto.getDiscountType())) {
+                Product product = getProduct(discountRequestDto);
+                Discount discount = Discount.create(
+                    product,
+                    discountRequestDto.getName(),
+                    discountRequestDto.getDiscount(),
+                    discountRequestDto.getStartedAt(),
+                    discountRequestDto.getExpiredAt());
                 product.changeDiscount(discountRequestDto.getDiscount());
                 productRepository.save(product);
+                discountRepository.save(discount);
+                return discount.getId();
             }
-            discountRepository.save(discount);
-            return discount.getId();
+
+            if (DiscountType.CATEGORY_DISCOUNT.equals(discountRequestDto.getDiscountType())) {
+                Category category = getCategory(discountRequestDto);
+                Discount discount = Discount.create(
+                    category,
+                    discountRequestDto.getName(),
+                    discountRequestDto.getDiscount(),
+                    discountRequestDto.getStartedAt(),
+                    discountRequestDto.getExpiredAt());
+                List<Product> products = categoryRepository.findProductsByCategoryId(category.getId());
+                for (Product product : products) {
+                    product.changeDiscount(discountRequestDto.getDiscount());
+                    productRepository.save(product);
+                }
+                discountRepository.save(discount);
+                return discount.getId();
+            }
         }
         throw new KkaKkaException();
     }
@@ -77,8 +78,24 @@ public class DiscountService {
     public void deleteDiscount(Long discountId) {
         Discount discount = discountRepository.findById(discountId)
             .orElseThrow(KkaKkaException::new);
-        discount.deleteDiscount();
-        discountRepository.save(discount);
+        if (DiscountType.PRODUCT_DISCOUNT.equals(discount.getDiscountType())) {
+            Product product = productRepository.findById(discount.getProductId())
+                .orElseThrow(KkaKkaException::new);
+            discount.deleteDiscount();
+            product.deleteDiscount();
+            discountRepository.save(discount);
+            productRepository.save(product);
+        }
+        if (DiscountType.CATEGORY_DISCOUNT.equals(discount.getDiscountType())) {
+            List<Product> products = categoryRepository.findProductsByCategoryId(
+                discount.getCategoryId());
+            for (Product product : products) {
+                product.deleteDiscount();
+                productRepository.save(product);
+            }
+            discount.deleteDiscount();
+            discountRepository.save(discount);
+        }
     }
 
     /* 할인 조회 */
